test(dead_code): split test_dead_code.py into a themed package
Break the 1713-line tests/unit/test_dead_code.py monolith into a
tests/unit/dead_code/ package: shared graph-builder/date helpers in
_helpers.py, tests grouped by scenario (reachability, unused exports,
unused internal, interfaces, decorators/dynamic, confidence, zombie
packages, whitelist/safety, alembic, web factory). Test bodies moved
verbatim; same 47 tests collected and passing.
