--- tests/unit/test_dead_code.py
@@ -1,1713 +1,0 @@
-"""Unit tests for DeadCodeAnalyzer."""
-
-from __future__ import annotations
-
-from datetime import UTC, datetime, timedelta
-
-import networkx as nx
-import pytest
-
-from repowise.core.analysis.dead_code import (
-    DeadCodeAnalyzer,
-    DeadCodeKind,
-)
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-
-def _now() -> datetime:
-    return datetime.now(UTC)
-
-
-def _old_date(days: int = 365) -> datetime:
-    """Return a datetime `days` ago (timezone-aware)."""
-    return _now() - timedelta(days=days)
-
-
-def _build_graph(
-    nodes: dict[str, dict],
-    edges: list[tuple[str, str]] | None = None,
-) -> nx.DiGraph:
-    """Create a DiGraph with the given node attributes and edges.
-
-    ``nodes`` maps node-id to its attribute dict.  If a node has a
-    ``symbols`` list, each entry is promoted to a proper symbol node
-    connected to the file via a ``defines`` edge (matching the real
-    GraphBuilder layout).
-    ``edges`` is a list of (src, dst) pairs; edge data can be
-    supplied as a 3-tuple (src, dst, data_dict).
-    """
-    g = nx.DiGraph()
-    for name, attrs in nodes.items():
-        attrs.setdefault("language", "python")
-        # Extract symbols before adding the file node
-        sym_list = attrs.pop("symbols", [])
-        g.add_node(name, **attrs)
-        # Create symbol nodes + defines edges (mirrors GraphBuilder.add_file)
-        for sym in sym_list:
-            sym_id = f"{name}::{sym['name']}"
-            g.add_node(
-                sym_id,
-                node_type="symbol",
-                file_path=name,
-                **sym,
-            )
-            g.add_edge(name, sym_id, edge_type="defines")
-    for edge in edges or []:
-        if len(edge) == 3:
-            g.add_edge(edge[0], edge[1], **(edge[2]))
-        else:
-            g.add_edge(edge[0], edge[1])
-    return g
-
-
-# ---------------------------------------------------------------------------
-# 1. test_unreachable_file_detected
-# ---------------------------------------------------------------------------
-
-
-def test_unreachable_file_detected():
-    """A file with in_degree=0, not an entry point, should be flagged as unreachable."""
-    g = _build_graph(
-        nodes={
-            "pkg/orphan.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 5,
-                "symbols": [],
-            },
-            "pkg/main.py": {
-                "is_entry_point": True,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 10,
-                "symbols": [],
-            },
-        },
-        edges=[],  # orphan.py has in_degree=0
-    )
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze({"detect_unused_exports": False, "detect_zombie_packages": False})
-
-    unreachable = [f for f in report.findings if f.kind == DeadCodeKind.UNREACHABLE_FILE]
-    paths = [f.file_path for f in unreachable]
-    assert "pkg/orphan.py" in paths
-
-
-# ---------------------------------------------------------------------------
-# 2. test_entry_point_not_flagged
-# ---------------------------------------------------------------------------
-
-
-def test_entry_point_not_flagged():
-    """A file marked as is_entry_point=True should NOT be flagged even with in_degree=0."""
-    g = _build_graph(
-        nodes={
-            "pkg/main.py": {
-                "is_entry_point": True,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 10,
-                "symbols": [],
-            },
-        },
-    )
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze({"detect_unused_exports": False, "detect_zombie_packages": False})
-
-    assert all(f.file_path != "pkg/main.py" for f in report.findings)
-
-
-# ---------------------------------------------------------------------------
-# 3. test_test_files_excluded
-# ---------------------------------------------------------------------------
-
-
-def test_test_files_excluded():
-    """A test file (is_test=True) with in_degree=0 should NOT be flagged."""
-    g = _build_graph(
-        nodes={
-            "tests/test_something.py": {
-                "is_entry_point": False,
-                "is_test": True,
-                "is_api_contract": False,
-                "symbol_count": 8,
-                "symbols": [],
-            },
-        },
-    )
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze({"detect_unused_exports": False, "detect_zombie_packages": False})
-
-    assert all(f.file_path != "tests/test_something.py" for f in report.findings)
-
-
-# ---------------------------------------------------------------------------
-# 4. test_unused_export_detected
-# ---------------------------------------------------------------------------
-
-
-def test_unused_export_detected():
-    """A public symbol with no importers should be flagged as unused export."""
-    g = _build_graph(
-        nodes={
-            "pkg/utils.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 3,
-                "symbols": [
-                    {
-                        "name": "helper_func",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 1,
-                        "end_line": 10,
-                        "complexity_estimate": 2,
-                    },
-                ],
-            },
-            "pkg/main.py": {
-                "is_entry_point": True,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 10,
-                "symbols": [],
-            },
-        },
-        # main.py imports utils.py but does NOT import helper_func by name
-        edges=[("pkg/main.py", "pkg/utils.py", {"imported_names": ["other_func"]})],
-    )
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_zombie_packages": False,
-        }
-    )
-
-    unused = [f for f in report.findings if f.kind == DeadCodeKind.UNUSED_EXPORT]
-    sym_names = [f.symbol_name for f in unused]
-    assert "helper_func" in sym_names
-
-
-def test_entry_point_names_never_flagged():
-    """Runtime entry points (wWinMain, TEST_METHOD, DllInstall…) are
-    never flagged regardless of incoming-edge counts. The Win32 / COM /
-    MSTest runners reach them by mechanism, not by a `using` import."""
-    g = _build_graph(
-        nodes={
-            "src/main.cpp": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 3,
-                "language": "cpp",
-                "symbols": [
-                    {
-                        "name": "wWinMain",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 1, "end_line": 10,
-                        "complexity_estimate": 1,
-                    },
-                    {
-                        "name": "TEST_METHOD",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 12, "end_line": 20,
-                        "complexity_estimate": 1,
-                    },
-                ],
-            },
-        },
-        edges=[],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {"detect_unreachable_files": False, "detect_zombie_packages": False}
-    )
-    names = [f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_EXPORT]
-    assert "wWinMain" not in names
-    assert "TEST_METHOD" not in names
-
-
-def test_unused_export_skipped_when_symbol_has_incoming_calls():
-    """A public symbol with no file-level importers but at least one
-    incoming ``calls`` edge on the symbol itself is still in use —
-    typical for C++ intra-file helpers and ``Foo::method`` qualified
-    definitions reached via call resolution rather than headers."""
-    g = _build_graph(
-        nodes={
-            "pkg/utils.cpp": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 2,
-                "language": "cpp",
-                "symbols": [
-                    {
-                        "name": "helper",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 1,
-                        "end_line": 10,
-                        "complexity_estimate": 1,
-                    },
-                    {
-                        "name": "caller",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 12,
-                        "end_line": 20,
-                        "complexity_estimate": 1,
-                    },
-                ],
-            },
-        },
-        edges=[
-            ("pkg/utils.cpp::caller", "pkg/utils.cpp::helper", {"edge_type": "calls"}),
-        ],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_zombie_packages": False,
-        }
-    )
-    sym_names = [f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_EXPORT]
-    # ``helper`` is called by ``caller`` — must not be flagged.
-    assert "helper" not in sym_names
-    # ``caller`` has no callers — still flagged.
-    assert "caller" in sym_names
-
-
-def test_interface_without_implementor_demoted_below_safe_threshold():
-    """Public interfaces with no incoming ``implements`` edges have
-    their unused-export confidence clamped below 0.7 so the demo
-    doesn't ship them as confident dead code. Implementor detection
-    is heuristic — absence is missing-signal, not evidence-of-absence.
-    """
-    g = _build_graph(
-        nodes={
-            "src/IBasketService.cs": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 1,
-                "symbols": [
-                    {
-                        "name": "IBasketService",
-                        "kind": "interface",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 1,
-                        "end_line": 5,
-                        "complexity_estimate": 1,
-                        "language": "csharp",
-                    },
-                ],
-            },
-        },
-        edges=[],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze({
-        "detect_unreachable_files": False,
-        "detect_zombie_packages": False,
-        "min_confidence": 0.0,
-    })
-    interface_findings = [
-        f for f in report.findings
-        if f.kind == DeadCodeKind.UNUSED_EXPORT and f.symbol_name == "IBasketService"
-    ]
-    # The finding may still surface, but never at safe-to-delete confidence.
-    for f in interface_findings:
-        assert f.confidence <= 0.4, f"interface flagged at unsafe confidence {f.confidence}"
-        assert f.safe_to_delete is False
-
-
-def test_com_contract_method_demoted_below_safe_threshold():
-    """``QueryInterface`` / ``AddRef`` / ``Release`` are reached through
-    a native COM vtable — never via a static caller. The analyzer must
-    not flag them as confident dead code.
-    """
-    g = _build_graph(
-        nodes={
-            "src/CoFoo.cpp": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 3,
-                "symbols": [
-                    {
-                        "name": "QueryInterface",
-                        "kind": "method",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 10,
-                        "end_line": 25,
-                        "complexity_estimate": 1,
-                        "language": "cpp",
-                    },
-                    {
-                        "name": "AddRef",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 27,
-                        "end_line": 30,
-                        "complexity_estimate": 1,
-                        "language": "cpp",
-                    },
-                    {
-                        "name": "Release",
-                        "kind": "method",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 32,
-                        "end_line": 40,
-                        "complexity_estimate": 1,
-                        "language": "cpp",
-                    },
-                ],
-            },
-        },
-        edges=[],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze({
-        "detect_unreachable_files": False,
-        "detect_zombie_packages": False,
-        "min_confidence": 0.0,
-    })
-    com_findings = [
-        f for f in report.findings
-        if f.symbol_name in {"QueryInterface", "AddRef", "Release"}
-    ]
-    for f in com_findings:
-        assert f.confidence <= 0.4, f"COM method flagged at unsafe confidence {f.confidence}"
-        assert f.safe_to_delete is False
-
-
-def test_release_in_non_com_language_not_clamped():
-    """A free function named ``Release`` in TypeScript/Python is *not* a
-    COM contract method — the contract-method clamp must not apply.
-    """
-    g = _build_graph(
-        nodes={
-            "src/foo.ts": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 1,
-                "symbols": [
-                    {
-                        "name": "Release",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 1,
-                        "end_line": 5,
-                        "complexity_estimate": 1,
-                        "language": "typescript",
-                    },
-                ],
-            },
-        },
-        edges=[],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze({
-        "detect_unreachable_files": False,
-        "detect_zombie_packages": False,
-        "min_confidence": 0.0,
-    })
-    # TS ``function`` kind is universally skipped from unused-export
-    # detection (functions surface as variables in JS-style modules),
-    # so the strongest assertion is "no spurious COM clamp masked the
-    # real behaviour". The function should still be detectable somehow
-    # downstream; here we just confirm the analyzer didn't crash and
-    # didn't ship it under contract-method semantics.
-    com_findings = [f for f in report.findings if f.symbol_name == "Release"]
-    for f in com_findings:
-        # If it surfaces at all, it must do so under the *normal*
-        # rule — not under the COM clamp specifically. The normal
-        # rule yields 0.7 / 1.0 depending on file-level importers.
-        assert f.confidence >= 0.7 or f.confidence < 0.4
-
-
-# ---------------------------------------------------------------------------
-# 5. test_framework_decorator_excluded
-# ---------------------------------------------------------------------------
-
-
-def test_framework_decorator_excluded():
-    """A symbol decorated with pytest.fixture should NOT be flagged."""
-    g = _build_graph(
-        nodes={
-            "pkg/conftest.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 2,
-                "symbols": [
-                    {
-                        "name": "db_session",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": ["pytest.fixture"],
-                        "start_line": 1,
-                        "end_line": 15,
-                        "complexity_estimate": 1,
-                    },
-                ],
-            },
-        },
-    )
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_zombie_packages": False,
-        }
-    )
-
-    sym_names = [f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_EXPORT]
-    assert "db_session" not in sym_names
-
-
-# ---------------------------------------------------------------------------
-# 6. test_dynamic_pattern_excluded
-# ---------------------------------------------------------------------------
-
-
-def test_dynamic_pattern_excluded():
-    """A symbol matching '*Handler' dynamic pattern should NOT be flagged as unused."""
-    g = _build_graph(
-        nodes={
-            "pkg/events.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 2,
-                "symbols": [
-                    {
-                        "name": "EventHandler",
-                        "kind": "class",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 1,
-                        "end_line": 40,
-                        "complexity_estimate": 3,
-                    },
-                ],
-            },
-        },
-    )
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_zombie_packages": False,
-        }
-    )
-
-    sym_names = [f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_EXPORT]
-    assert "EventHandler" not in sym_names
-
-
-# ---------------------------------------------------------------------------
-# 6b. test_nested_function_not_flagged
-# ---------------------------------------------------------------------------
-
-
-def test_nested_function_not_flagged():
-    """Symbols defined inside another function's body cannot be imported by
-    name and must not be flagged as unused exports."""
-    g = _build_graph(
-        nodes={
-            "pkg/server.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 2,
-                "symbols": [
-                    {
-                        "name": "chat",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 10,
-                        "end_line": 80,
-                    },
-                    # Nested closure / inner generator — line range strictly
-                    # contained inside ``chat``.
-                    {
-                        "name": "event_stream",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 25,
-                        "end_line": 60,
-                    },
-                ],
-            },
-        },
-    )
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_zombie_packages": False,
-        }
-    )
-    sym_names = [f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_EXPORT]
-    assert "event_stream" not in sym_names
-
-
-# ---------------------------------------------------------------------------
-# 6c. test_at_prefixed_decorator_excluded
-# ---------------------------------------------------------------------------
-
-
-def test_at_prefixed_decorator_excluded():
-    """Decorators stored with the leading '@' (as the parser emits them)
-    must still match the framework-decorator whitelist."""
-    g = _build_graph(
-        nodes={
-            "pkg/routes.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 2,
-                "symbols": [
-                    {
-                        "name": "list_items",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": ["@router.get(\"/items\")"],
-                        "start_line": 1,
-                        "end_line": 8,
-                    },
-                    {
-                        "name": "lifespan",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": ["@asynccontextmanager"],
-                        "start_line": 10,
-                        "end_line": 20,
-                    },
-                ],
-            },
-        },
-    )
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_zombie_packages": False,
-        }
-    )
-    sym_names = [f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_EXPORT]
-    assert "list_items" not in sym_names
-    assert "lifespan" not in sym_names
-
-
-# ---------------------------------------------------------------------------
-# 7. test_confidence_low_for_recent_files
-# ---------------------------------------------------------------------------
-
-
-def test_confidence_low_for_recent_files():
-    """Unreachable file with commit_count_90d > 0 should have confidence 0.4."""
-    g = _build_graph(
-        nodes={
-            "pkg/recent.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 5,
-                "symbols": [],
-            },
-        },
-    )
-
-    git_meta = {
-        "pkg/recent.py": {
-            "commit_count_90d": 3,
-            "last_commit_at": _now() - timedelta(days=10),
-            "age_days": 100,
-            "primary_owner_name": "dev@example.com",
-        },
-    }
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map=git_meta)
-    report = analyzer.analyze(
-        {
-            "detect_unused_exports": False,
-            "detect_zombie_packages": False,
-            "min_confidence": 0.0,
-        }
-    )
-
-    findings = [f for f in report.findings if f.file_path == "pkg/recent.py"]
-    assert len(findings) == 1
-    assert findings[0].confidence == pytest.approx(0.4)
-
-
-# ---------------------------------------------------------------------------
-# 8. test_confidence_high_for_stale_unreachable
-# ---------------------------------------------------------------------------
-
-
-def test_confidence_high_for_stale_unreachable():
-    """Unreachable file with no commits in 90d and last commit > 6 months ago -> confidence 1.0."""
-    g = _build_graph(
-        nodes={
-            "pkg/stale.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 5,
-                "symbols": [],
-            },
-        },
-    )
-
-    git_meta = {
-        "pkg/stale.py": {
-            "commit_count_90d": 0,
-            "last_commit_at": _old_date(days=365),
-            "age_days": 730,
-            "primary_owner_name": "dev@example.com",
-        },
-    }
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map=git_meta)
-    report = analyzer.analyze(
-        {
-            "detect_unused_exports": False,
-            "detect_zombie_packages": False,
-        }
-    )
-
-    findings = [f for f in report.findings if f.file_path == "pkg/stale.py"]
-    assert len(findings) == 1
-    assert findings[0].confidence == pytest.approx(1.0)
-
-
-# ---------------------------------------------------------------------------
-# 9. test_zombie_package_detected
-# ---------------------------------------------------------------------------
-
-
-def test_zombie_package_detected():
-    """A package with no incoming inter-package imports should be flagged as zombie."""
-    g = _build_graph(
-        nodes={
-            "pkgA/mod1.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 5,
-                "symbols": [],
-            },
-            "pkgA/mod2.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 3,
-                "symbols": [],
-            },
-            "pkgB/mod1.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 7,
-                "symbols": [],
-            },
-        },
-        edges=[
-            # pkgA/mod1 imports from pkgA/mod2 (intra-package only)
-            ("pkgA/mod1.py", "pkgA/mod2.py"),
-            # pkgB has no inter-package importers either, but we focus on pkgA
-            # having NO imports from pkgB -> pkgA
-        ],
-    )
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_unused_exports": False,
-            "min_confidence": 0.0,
-        }
-    )
-
-    zombie = [f for f in report.findings if f.kind == DeadCodeKind.ZOMBIE_PACKAGE]
-    pkgs = [f.package for f in zombie]
-    # Both pkgA and pkgB are zombie since neither has inter-package importers
-    assert "pkgA" in pkgs
-    assert "pkgB" in pkgs
-
-
-def test_framework_anchor_counts_as_cross_package_importer():
-    """``framework:`` synthetic predecessors should rescue a package from zombie status.
-
-    ``external:`` predecessors don't count (third-party imports), but
-    ``framework:`` anchors model real framework-mediated loading
-    (e.g. TYPO3 core loading every ``Configuration/*.php``) and should.
-    """
-    g = _build_graph(
-        nodes={
-            "Configuration/foo.php": {"is_entry_point": False, "symbol_count": 1, "symbols": []},
-            "src/main.php": {"is_entry_point": False, "symbol_count": 1, "symbols": []},
-        },
-    )
-    g.add_node("framework:typo3-core", language="external")
-    g.add_edge("framework:typo3-core", "Configuration/foo.php", edge_type="framework")
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_unused_exports": False,
-            "min_confidence": 0.0,
-        }
-    )
-
-    zombie_pkgs = [
-        f.package for f in report.findings if f.kind == DeadCodeKind.ZOMBIE_PACKAGE
-    ]
-    assert "Configuration" not in zombie_pkgs, (
-        "framework: predecessors should count as cross-package importers"
-    )
-
-
-def test_framework_anchor_node_not_flagged_as_unreachable():
-    """A ``framework:`` synthetic node must not itself be flagged as dead."""
-    g = _build_graph(
-        nodes={
-            "ext_localconf.php": {"is_entry_point": False, "symbol_count": 1, "symbols": []},
-        },
-    )
-    g.add_node("framework:typo3-core", language="external")
-    g.add_edge("framework:typo3-core", "ext_localconf.php", edge_type="framework")
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze({"min_confidence": 0.0})
-
-    paths = [f.file_path for f in report.findings]
-    assert "framework:typo3-core" not in paths
-    # ext_localconf.php should also not be flagged: it has a framework: predecessor.
-    assert "ext_localconf.php" not in paths
-
-
-# ---------------------------------------------------------------------------
-# 10. test_whitelist_respected
-# ---------------------------------------------------------------------------
-
-
-def test_whitelist_respected():
-    """A file in the whitelist should NOT be flagged even if it is unreachable."""
-    g = _build_graph(
-        nodes={
-            "pkg/legacy.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 20,
-                "symbols": [],
-            },
-        },
-    )
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unused_exports": False,
-            "detect_zombie_packages": False,
-            "whitelist": ["pkg/legacy.py"],
-        }
-    )
-
-    assert all(f.file_path != "pkg/legacy.py" for f in report.findings)
-
-
-# ---------------------------------------------------------------------------
-# 11. test_safe_to_delete_conservative
-# ---------------------------------------------------------------------------
-
-
-def test_safe_to_delete_conservative():
-    """safe_to_delete is True only when confidence >= 0.7 AND file does not match dynamic patterns."""
-    g = _build_graph(
-        nodes={
-            # High confidence, no dynamic pattern match -> safe
-            "pkg/old_unused.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 5,
-                "symbols": [],
-            },
-            # High confidence, but file stem matches *Handler -> NOT safe
-            "pkg/RequestHandler.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 5,
-                "symbols": [],
-            },
-            # Low confidence (recently touched) -> NOT safe
-            "pkg/fresh.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 5,
-                "symbols": [],
-            },
-        },
-    )
-
-    git_meta = {
-        "pkg/old_unused.py": {
-            "commit_count_90d": 0,
-            "last_commit_at": _old_date(days=365),
-            "age_days": 500,
-        },
-        "pkg/RequestHandler.py": {
-            "commit_count_90d": 0,
-            "last_commit_at": _old_date(days=365),
-            "age_days": 500,
-        },
-        "pkg/fresh.py": {
-            "commit_count_90d": 5,
-            "last_commit_at": _now() - timedelta(days=3),
-            "age_days": 60,
-        },
-    }
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map=git_meta)
-    report = analyzer.analyze(
-        {
-            "detect_unused_exports": False,
-            "detect_zombie_packages": False,
-            "min_confidence": 0.0,
-        }
-    )
-
-    by_path = {f.file_path: f for f in report.findings}
-    # High confidence + no dynamic pattern -> safe
-    assert by_path["pkg/old_unused.py"].safe_to_delete is True
-    # High confidence but matches *Handler -> not safe
-    assert by_path["pkg/RequestHandler.py"].safe_to_delete is False
-    # Low confidence (0.4) -> not safe
-    assert by_path["pkg/fresh.py"].safe_to_delete is False
-
-
-# ---------------------------------------------------------------------------
-# 12. test_report_deletable_lines_sum
-# ---------------------------------------------------------------------------
-
-
-def test_report_deletable_lines_sum():
-    """report.deletable_lines should equal the sum of lines for safe_to_delete findings."""
-    g = _build_graph(
-        nodes={
-            "pkg/dead1.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 10,  # lines = 10 * 10 = 100
-                "symbols": [],
-            },
-            "pkg/dead2.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 20,  # lines = 20 * 10 = 200
-                "symbols": [],
-            },
-            "pkg/alive.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 15,  # lines = 15 * 10 = 150, but NOT safe
-                "symbols": [],
-            },
-        },
-    )
-
-    git_meta = {
-        "pkg/dead1.py": {
-            "commit_count_90d": 0,
-            "last_commit_at": _old_date(days=365),
-            "age_days": 400,
-        },
-        "pkg/dead2.py": {
-            "commit_count_90d": 0,
-            "last_commit_at": _old_date(days=365),
-            "age_days": 400,
-        },
-        # Recently touched -> confidence 0.4, safe_to_delete=False
-        "pkg/alive.py": {
-            "commit_count_90d": 5,
-            "last_commit_at": _now() - timedelta(days=2),
-            "age_days": 60,
-        },
-    }
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map=git_meta)
-    report = analyzer.analyze(
-        {
-            "detect_unused_exports": False,
-            "detect_zombie_packages": False,
-            "min_confidence": 0.0,
-        }
-    )
-
-    safe_findings = [f for f in report.findings if f.safe_to_delete]
-    expected_lines = sum(f.lines for f in safe_findings)
-    assert report.deletable_lines == expected_lines
-    # Verify that the safe findings include the two stale files
-    safe_paths = {f.file_path for f in safe_findings}
-    assert "pkg/dead1.py" in safe_paths
-    assert "pkg/dead2.py" in safe_paths
-    assert "pkg/alive.py" not in safe_paths
-    # Verify the actual sum
-    assert report.deletable_lines == 100 + 200
-
-
-# ---------------------------------------------------------------------------
-# 13. test_dynamic_edge_clamps_unreachable_confidence
-# ---------------------------------------------------------------------------
-
-
-def test_dynamic_edge_clamps_unreachable_confidence():
-    """A file in a package that has dynamic graph edges should have confidence clamped to 0.4."""
-    g = _build_graph(
-        nodes={
-            "pkg/orphan.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 5,
-                "symbols": [],
-            },
-            "pkg/dispatcher.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 3,
-                "symbols": [],
-            },
-            "pkg/handler.py": {
-                "is_entry_point": True,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 3,
-                "symbols": [],
-            },
-        },
-        edges=[
-            ("pkg/dispatcher.py", "pkg/handler.py", {"edge_type": "dynamic"}),
-        ],
-    )
-
-    git_meta = {
-        "pkg/orphan.py": {
-            "commit_count_90d": 0,
-            "last_commit_at": _old_date(days=400),
-            "age_days": 500,
-        },
-    }
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map=git_meta)
-    report = analyzer.analyze(
-        {
-            "detect_unused_exports": False,
-            "detect_zombie_packages": False,
-            "min_confidence": 0.0,
-        }
-    )
-
-    findings = [f for f in report.findings if f.file_path == "pkg/orphan.py"]
-    assert len(findings) == 1
-    assert findings[0].confidence == pytest.approx(0.4)
-    assert any("dynamic" in e.lower() for e in findings[0].evidence)
-
-
-# ---------------------------------------------------------------------------
-# 14. test_find_dynamic_edge_files_handles_subtypes
-# ---------------------------------------------------------------------------
-
-
-def test_find_dynamic_edge_files_handles_subtypes():
-    """find_dynamic_edge_files picks up edges with dynamic_* sub-type prefixes."""
-    from repowise.core.analysis.dead_code.dynamic_markers import find_dynamic_edge_files
-
-    g = nx.DiGraph()
-    g.add_node("a.py")
-    g.add_node("b.py")
-    g.add_node("c.py")
-    g.add_node("d.py")
-    g.add_node("external:something")
-    g.add_edge("a.py", "b.py", edge_type="dynamic_uses")
-    g.add_edge("c.py", "d.py", edge_type="dynamic_imports")
-    g.add_edge("a.py", "external:something", edge_type="dynamic")
-    # Non-dynamic edges should not contribute
-    g.add_edge("b.py", "c.py", edge_type="calls")
-
-    files = find_dynamic_edge_files(g)
-    assert {"a.py", "b.py", "c.py", "d.py"} <= files
-    assert "external:something" not in files
-
-
-# ---------------------------------------------------------------------------
-# 15. test_dynamic_markers_per_language_coverage
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.parametrize(
-    "suffix,expected_markers",
-    [
-        (".go", ("plugin.Open(", "reflect.New(", "reflect.TypeOf(", "reflect.ValueOf(")),
-        (
-            ".rb",
-            ("autoload ", "const_get(", "Object.send(", "Kernel.const_get(", ".public_send("),
-        ),
-        (
-            ".php",
-            (
-                "class_exists(",
-                "call_user_func(",
-                "call_user_func_array(",
-                "new $",
-                "ReflectionClass(",
-            ),
-        ),
-        (
-            ".kt",
-            ("Class.forName(", "ServiceLoader.load(", "KClass.createInstance(", "::class.java"),
-        ),
-        (
-            ".swift",
-            ("NSClassFromString(", "Selector(", "#selector(", "NSStringFromClass("),
-        ),
-        (".scala", ("Class.forName(", "runtimeMirror(", "reflect.runtime")),
-    ],
-)
-def test_dynamic_markers_per_language_coverage(suffix, expected_markers):
-    """Each language's marker tuple must contain the documented Phase 2 entries."""
-    from repowise.core.analysis.dead_code.dynamic_markers import _DYNAMIC_IMPORT_MARKERS
-
-    markers = _DYNAMIC_IMPORT_MARKERS.get(suffix, ())
-    for expected in expected_markers:
-        assert expected in markers, f"missing marker {expected!r} for {suffix}"
-
-
-# ---------------------------------------------------------------------------
-# 16. test_unused_internal_default_on
-# ---------------------------------------------------------------------------
-
-
-def test_unused_internal_default_on():
-    """detect_unused_internals defaults to True; private symbols with no callers are flagged."""
-    g = _build_graph(
-        nodes={
-            "pkg/utils.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 1,
-                "symbols": [
-                    {
-                        "name": "_helper",
-                        "kind": "function",
-                        "visibility": "private",
-                        "decorators": [],
-                        "start_line": 1,
-                        "end_line": 12,
-                        "complexity_estimate": 1,
-                    },
-                ],
-            },
-        },
-    )
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_unused_exports": False,
-            "detect_zombie_packages": False,
-            "min_confidence": 0.0,
-        }
-    )
-
-    internals = [f for f in report.findings if f.kind == DeadCodeKind.UNUSED_INTERNAL]
-    assert any(f.symbol_name == "_helper" for f in internals)
-    assert all(f.safe_to_delete is False for f in internals)
-
-
-# ---------------------------------------------------------------------------
-# 17. test_unused_internal_explicit_opt_out
-# ---------------------------------------------------------------------------
-
-
-def test_unused_internal_explicit_opt_out():
-    """Setting detect_unused_internals=False disables the detector."""
-    g = _build_graph(
-        nodes={
-            "pkg/utils.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 1,
-                "symbols": [
-                    {
-                        "name": "_helper",
-                        "kind": "function",
-                        "visibility": "private",
-                        "decorators": [],
-                        "start_line": 1,
-                        "end_line": 12,
-                        "complexity_estimate": 1,
-                    },
-                ],
-            },
-        },
-    )
-
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_unused_exports": False,
-            "detect_zombie_packages": False,
-            "detect_unused_internals": False,
-            "min_confidence": 0.0,
-        }
-    )
-
-    internals = [f for f in report.findings if f.kind == DeadCodeKind.UNUSED_INTERNAL]
-    assert internals == []
-
-
-# ---------------------------------------------------------------------------
-# D3. Alembic migration scripts are never flagged
-# ---------------------------------------------------------------------------
-
-
-def test_alembic_versions_never_flagged_as_unreachable():
-    """Files under <root>/alembic/versions/*.py are loaded reflectively by
-    Alembic at runtime — they have no static importer and must not be
-    surfaced as unreachable on any Alembic-using repo (generic Python
-    convention)."""
-    g = _build_graph(
-        nodes={
-            "myapp/alembic/versions/0001_init.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 2,
-                "symbols": [],
-            },
-            "myapp/alembic/versions/0002_add_users.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 2,
-                "symbols": [],
-            },
-        },
-        edges=[],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {"detect_unused_exports": False, "detect_zombie_packages": False}
-    )
-    paths = {f.file_path for f in report.findings}
-    assert "myapp/alembic/versions/0001_init.py" not in paths
-    assert "myapp/alembic/versions/0002_add_users.py" not in paths
-
-
-def test_alembic_upgrade_downgrade_never_flagged_as_unused():
-    """``upgrade()`` / ``downgrade()`` in Alembic migration scripts are
-    called via reflection by the Alembic runner — the file pattern
-    exemption (``*/alembic/versions/*.py``) already covers them, so the
-    symbols inside must not surface as unused exports either."""
-    g = _build_graph(
-        nodes={
-            "myapp/alembic/versions/0001_init.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 2,
-                "symbols": [
-                    {
-                        "name": "upgrade",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 1, "end_line": 5,
-                        "complexity_estimate": 1,
-                    },
-                    {
-                        "name": "downgrade",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 7, "end_line": 11,
-                        "complexity_estimate": 1,
-                    },
-                ],
-            },
-        },
-        edges=[],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {"detect_unreachable_files": False, "detect_zombie_packages": False}
-    )
-    names = {f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_EXPORT}
-    assert "upgrade" not in names
-    assert "downgrade" not in names
-
-
-# ---------------------------------------------------------------------------
-# D4. Click @<group>.command(...) decorator suffix detection
-# ---------------------------------------------------------------------------
-
-
-def test_click_group_command_decorator_excluded():
-    """Subcommands registered on a locally-named Click ``Group`` are
-    decorated as e.g. ``@my_cli.command("add")``. The receiver name is
-    project-local, so the matcher needs to recognise the ``.command``
-    suffix rather than a hard-coded ``click.command`` prefix. Same shape
-    covers Typer (``.command``/``.callback``) and any user-named
-    dispatcher."""
-    g = _build_graph(
-        nodes={
-            "pkg/cli.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 3,
-                "symbols": [
-                    {
-                        "name": "decision_add",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": ["decision_group.command"],
-                        "start_line": 1, "end_line": 10,
-                        "complexity_estimate": 1,
-                    },
-                    {
-                        "name": "decision_list",
-                        "kind": "function",
-                        "visibility": "public",
-                        # Real-world capture sometimes includes the call
-                        # arguments — the matcher must strip them.
-                        "decorators": ['my_cli.group("decision")'],
-                        "start_line": 12, "end_line": 20,
-                        "complexity_estimate": 1,
-                    },
-                    {
-                        "name": "after_cmd",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": ["app.callback"],
-                        "start_line": 22, "end_line": 28,
-                        "complexity_estimate": 1,
-                    },
-                ],
-            },
-        },
-        edges=[],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {"detect_unreachable_files": False, "detect_zombie_packages": False}
-    )
-    names = {f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_EXPORT}
-    assert "decision_add" not in names
-    assert "decision_list" not in names
-    assert "after_cmd" not in names
-
-
-def test_unrelated_dot_command_function_still_flagged():
-    """A decorator suffix-match must not accidentally whitelist symbols
-    with unrelated decorators. ``@functools.lru_cache`` does not end with
-    ``.command/.group/.callback``, so a plain unused public helper with
-    that decorator is still surfaced."""
-    g = _build_graph(
-        nodes={
-            "pkg/helpers.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 1,
-                "symbols": [
-                    {
-                        "name": "expensive_lookup",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": ["functools.lru_cache"],
-                        "start_line": 1, "end_line": 20,
-                        "complexity_estimate": 1,
-                    },
-                ],
-            },
-        },
-        edges=[],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {"detect_unreachable_files": False, "detect_zombie_packages": False}
-    )
-    names = {f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_EXPORT}
-    assert "expensive_lookup" in names
-
-
-# ---------------------------------------------------------------------------
-# D7. unused_internal: imports-by-name into the symbol's file means used
-# ---------------------------------------------------------------------------
-
-
-def test_unused_internal_skipped_when_imported_by_name():
-    """A private helper imported by name into a sibling module (typical
-    dispatch-table pattern: ``HANDLERS = {"python": _extract_py, ...}``)
-    is reached at runtime via dict lookup — no direct ``calls`` edge
-    will exist, but the ``imports`` edge carries the symbol name. Such
-    helpers must not be flagged as unused internals."""
-    g = _build_graph(
-        nodes={
-            "pkg/python_handler.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 1,
-                "symbols": [
-                    {
-                        "name": "_extract_python",
-                        "kind": "function",
-                        "visibility": "private",
-                        "decorators": [],
-                        "start_line": 1, "end_line": 15,
-                        "complexity_estimate": 1,
-                    },
-                ],
-            },
-            "pkg/dispatch.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 0,
-                "symbols": [],
-            },
-        },
-        edges=[
-            (
-                "pkg/dispatch.py",
-                "pkg/python_handler.py",
-                {"edge_type": "imports", "imported_names": ["_extract_python"]},
-            ),
-        ],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_unused_exports": False,
-            "detect_zombie_packages": False,
-            "min_confidence": 0.0,
-        }
-    )
-    names = {f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_INTERNAL}
-    assert "_extract_python" not in names
-
-
-def test_unused_export_skipped_when_file_imported_as_namespace():
-    """A public symbol in a file that is imported by its module name
-    (``from . import cargo``) is a dispatch-table candidate — the
-    importer holds a binding to the module and can call any public
-    attribute via ``cargo.<attr>(...)``. We cannot tell statically
-    which attribute is used, so every public symbol in the file must
-    be treated as live."""
-    g = _build_graph(
-        nodes={
-            "pkg/external_systems/cargo.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 1,
-                "symbols": [
-                    {
-                        "name": "parse",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 1, "end_line": 12,
-                        "complexity_estimate": 1,
-                    },
-                ],
-            },
-            "pkg/external_systems/__init__.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 0,
-                "symbols": [],
-            },
-        },
-        edges=[
-            (
-                "pkg/external_systems/__init__.py",
-                "pkg/external_systems/cargo.py",
-                {"edge_type": "imports", "imported_names": ["cargo"]},
-            ),
-        ],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_unused_internals": False,
-            "detect_zombie_packages": False,
-            "min_confidence": 0.0,
-        }
-    )
-    names = {f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_EXPORT}
-    assert "parse" not in names
-
-
-def test_unused_export_still_flagged_when_namespace_name_differs():
-    """Negative test for the namespace-import rescue: if the importer
-    pulls a *different* sibling module, the public symbol's file is
-    still effectively orphan and the symbol must be flagged."""
-    g = _build_graph(
-        nodes={
-            "pkg/external_systems/cargo.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 1,
-                "symbols": [
-                    {
-                        "name": "parse",
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 1, "end_line": 12,
-                        "complexity_estimate": 1,
-                    },
-                ],
-            },
-            "pkg/external_systems/__init__.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 0,
-                "symbols": [],
-            },
-        },
-        edges=[
-            # importer pulled "npm" only — does NOT mention "cargo"
-            (
-                "pkg/external_systems/__init__.py",
-                "pkg/external_systems/cargo.py",
-                {"edge_type": "imports", "imported_names": ["npm"]},
-            ),
-        ],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_unused_internals": False,
-            "detect_zombie_packages": False,
-            "min_confidence": 0.0,
-        }
-    )
-    names = {f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_EXPORT}
-    assert "parse" in names
-
-
-def test_unused_export_not_rescued_for_index_stem():
-    """``index.ts`` is the barrel filename in JS/TS package layouts —
-    every workspace package's source root has one, and most imports of
-    a package land on it. We must not over-rescue every public export
-    of every ``index.ts`` just because a downstream importer happens
-    to use the literal name ``index``. Same applies to Python
-    ``__init__`` (already covered by the global never-flag list)."""
-    g = _build_graph(
-        nodes={
-            "pkg/ui/src/index.ts": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "language": "typescript",
-                "symbol_count": 1,
-                "symbols": [
-                    {
-                        "name": "stranded_export",
-                        "kind": "function",
-                        "visibility": "public",
-                        "language": "typescript",
-                        "decorators": [],
-                        "start_line": 1, "end_line": 5,
-                        "complexity_estimate": 1,
-                    },
-                ],
-            },
-            "pkg/web/src/page.ts": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "language": "typescript",
-                "symbol_count": 0,
-                "symbols": [],
-            },
-        },
-        edges=[
-            (
-                "pkg/web/src/page.ts",
-                "pkg/ui/src/index.ts",
-                {"edge_type": "imports", "imported_names": ["index"]},
-            ),
-        ],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_unused_internals": False,
-            "detect_zombie_packages": False,
-            "min_confidence": 0.0,
-        }
-    )
-    names = {f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_EXPORT}
-    assert "stranded_export" in names
-
-
-def test_unused_internal_still_flagged_when_imports_dont_carry_name():
-    """Sanity check: an ``imports`` edge that does NOT list the private
-    symbol's name (e.g. the importer pulls a different sibling symbol)
-    must not rescue the helper from the unused-internal pass."""
-    g = _build_graph(
-        nodes={
-            "pkg/helpers.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 2,
-                "symbols": [
-                    {
-                        "name": "_unused_helper",
-                        "kind": "function",
-                        "visibility": "private",
-                        "decorators": [],
-                        "start_line": 1, "end_line": 8,
-                        "complexity_estimate": 1,
-                    },
-                ],
-            },
-            "pkg/consumer.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 0,
-                "symbols": [],
-            },
-        },
-        edges=[
-            (
-                "pkg/consumer.py",
-                "pkg/helpers.py",
-                {"edge_type": "imports", "imported_names": ["something_else"]},
-            ),
-        ],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {
-            "detect_unreachable_files": False,
-            "detect_unused_exports": False,
-            "detect_zombie_packages": False,
-            "min_confidence": 0.0,
-        }
-    )
-    names = {f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_INTERNAL}
-    assert "_unused_helper" in names
-
-
-# ---------------------------------------------------------------------------
-# D10. WSGI / ASGI / app-factory names treated as entry points
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.parametrize(
-    "factory_name",
-    [
-        "create_app",
-        "make_app",
-        "application",
-        "asgi_app",
-        "wsgi_app",
-        "get_asgi_application",
-        "get_wsgi_application",
-    ],
-)
-def test_python_web_factory_not_flagged(factory_name):
-    """FastAPI / Flask / Tornado / aiohttp / Django entry symbols are
-    loaded by an external server via dotted-path string
-    (``module:create_app`` / ``module:application``) — no graph edge
-    exists from the launching server. These conventional names must be
-    in the entry-point allowlist so the unused-export pass skips them."""
-    g = _build_graph(
-        nodes={
-            "myapp/server.py": {
-                "is_entry_point": False,
-                "is_test": False,
-                "is_api_contract": False,
-                "symbol_count": 1,
-                "symbols": [
-                    {
-                        "name": factory_name,
-                        "kind": "function",
-                        "visibility": "public",
-                        "decorators": [],
-                        "start_line": 1, "end_line": 20,
-                        "complexity_estimate": 1,
-                    },
-                ],
-            },
-        },
-        edges=[],
-    )
-    analyzer = DeadCodeAnalyzer(g, git_meta_map={})
-    report = analyzer.analyze(
-        {"detect_unreachable_files": False, "detect_zombie_packages": False}
-    )
-    names = {f.symbol_name for f in report.findings if f.kind == DeadCodeKind.UNUSED_EXPORT}
-    assert factory_name not in names
